requirements.txt
# Core ML/Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning (for neural collaborative filtering)
torch==2.0.1
tensorflow==2.13.0

# Recommendation Systems
surprise==0.1
implicit==0.7.2

# API Framework
fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.0.3

# Data Validation & Processing
pandera==0.15.2
pyyaml==6.0

# Experiment Tracking
mlflow==2.5.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.65.0

# Visualization (optional, for notebooks)
matplotlib==3.7.2
seaborn==0.12.2